Convert MESA into python package
This converts MESA from a group of scripts into a python package. A
setup.py, MANIFEST, etc. have been added and the individual scripts are
now moved into a new mesa directory. Each of the individual scripts are
now called through a single entry point in __main__.py. Argument parsers
are added and consumed by each file having a add_parser() function and
uses the arguments with a run_with() function. Its important to set
certain options such as set_defaults() in add_parser correctly or it may
lead to subtle bugs.
This commit also uses black to provide automatic formatting of the code,
however most dead code has not yet been removed and will need to do so
in the future.